Rename lastMatch.py to last_match.py (OCD)
